# docs/swagger.yaml
basePath: /
definitions:
  request.CreateELogBookRequest:
    properties:
      deskripsi:
        type: string
      end_time:
        type: integer
      id_konsulen:
        type: integer
      jumlah:
        type: integer
      medical_record:
        type: string
      start_time:
        type: integer
      title:
        type: string
    type: object
...
  request.UpdateELogBookRequest:
    properties:
      deskripsi:
        type: string
      end_time:
        type: integer
      jumlah:
        type: integer
      medical_record:
        type: string
      start_time:
        type: integer
      title:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API yang digunakan untuk website SIPEKOM
  title: API Sipekom
  version: 0.0.1
paths:
...
  /api/elogbook/:
    get:
      description: get all ELogBook
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get all ELogBook [mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/accepted/{id_elogbook}:
    put:
      description: Approved ELogBook by ID.
      parameters:
      - description: ID ELogBook
        in: path
        name: id_elogbook
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "Approved ELogBook [mahasiswa\U0001F512, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/create:
    post:
      consumes:
      - application/json
      description: create new ELogBook.
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateELogBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "create ELogBook [konsulen\U0001F512, mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/delete/{id_elogbook}:
    delete:
      consumes:
      - application/json
      description: delete ELogBook by ID.
      parameters:
      - description: ID ELogBook
        in: path
        name: id_elogbook
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "delete ELogBook [konsulen\U0001F512, mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/get/{id_elogbook}:
    get:
      description: get ELogBook by id user.
      parameters:
      - description: ID Elogbook
        in: path
        name: id_elogbook
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get ELogBook [mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/rejected/{id_elogbook}:
    put:
      description: Approved ELogBook by ID.
      parameters:
      - description: ID ELogBook
        in: path
        name: id_elogbook
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "Approved ELogBook [mahasiswa\U0001F512, guest\U0001F512]."
      tags:
      - ELogBook
  /api/elogbook/update/{id_elogbook}:
    put:
      description: update ELogBook by ID.
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateELogBookRequest'
      - description: ID ELogBook
        in: path
        name: id_elogbook
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "update ELogBook [konsulen\U0001F512, mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - ELogBook
...
securityDefinitions:
  ApiKeyAuth:
    description: API Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"